requirements.txt
# 核心
torch>=2.2
numpy>=1.24
PyYAML>=6.0


# 服務
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6


# 訓練輔助
tqdm>=4.66
